feat(contest): Add components for contesr page
