+++ src/features/home/components/contest/Background.tsx
@@ -1,0 +1,42 @@
+import { useState, useEffect } from 'react';
+
+const BackgroundEffects = () => {
+  const [mousePosition, setMousePosition] = useState({ x: 0, y: 0 });
+
+  useEffect(() => {
+    const handleMouseMove = (e: MouseEvent) => {
+      setMousePosition({ x: e.clientX, y: e.clientY });
+    };
+    window.addEventListener('mousemove', handleMouseMove);
+    return () => window.removeEventListener('mousemove', handleMouseMove);
+  }, []);
+
+  return (
+    <>
+      <div className="fixed inset-0 overflow-hidden pointer-events-none">
+        {[...Array(50)].map((_, i) => (
+          <div
+            key={i}
+            className="absolute w-1 h-1 bg-blue-400 rounded-full animate-pulse"
+            style={{
+              left: `${Math.random() * 100}%`,
+              top: `${Math.random() * 100}%`,
+              animationDelay: `${Math.random() * 3}s`,
+              animationDuration: `${2 + Math.random() * 3}s`
+            }}
+          />
+        ))}
+      </div>
+      <div
+        className="fixed w-4 h-4 bg-gradient-to-r from-blue-400 to-blue-600 rounded-full pointer-events-none mix-blend-screen opacity-70 transition-all duration-100 ease-out z-50"
+        style={{
+          left: mousePosition.x - 8,
+          top: mousePosition.y - 8,
+          transform: 'translate3d(0, 0, 0)'
+        }}
+      />
+    </>
+  );
+};
+
+export default BackgroundEffects;
